Names on search bar and search button
Added name and type to search bar and name to search button in order to fetch data from the database.
Cooking, Alchemy, and Processing Calculators now all retrieve item images and item names  that are like the text entered in the search.
Still need to escape and check the text coming from user.