Added 3 more tables, and new Javascript files for functionality.
Added 3 tables, 1 is for the cost of the recipe, 1 is for the market price values of the recipes rewards, and the last table is for calculating the costs of each of the sub materials.
Several javascript files were added for the calculations of the costs for the sub materials when a number of the price is entered.