Rework of Javascript and some SASS
Went with a module approach to the javascript in my application.
One module for the recipeController, another for the UIController, and a final one that connects the two together.
The idea is to have the recipe controller configure all  the calculations, then have  the ui controller update the UI with the newly calculated variables.
The appController is their ways of communication.
Also edited more SASS for the tables and media queries.