Quantities, Costs, and Totals calculations now working.
Completed all Javascript implementation for the calculations for the recipes page.
Several functions to grab the initial quantities of the sub materials and proc rate. Then when the user enters more quantities of that  recipe,the quantities, the costs of the materials, and the profit from the rewards all update.
Added a check to see if the recipe rewards 2 recipe items or just 1, if just 1 we only need to calculate 1 profit row.